Add 1.21.8 support Addresses https://github.com/stefvanschie/IF/issues/2382
