+++ .github/workflows/maven.yml
@@ -127,7 +127,7 @@ jobs:
         java -jar BuildTools.jar --rev 1.21.3 --remapped --disable-java-check
         java -jar BuildTools.jar --rev 1.21.4 --remapped --disable-java-check
         java -jar BuildTools.jar --rev 1.21.5 --remapped --disable-java-check
-        java -jar BuildTools.jar --rev 1.21.7 --remapped --disable-java-check
+        java -jar BuildTools.jar --rev 1.21.8 --remapped --disable-java-check
         
         cd ../
     - name: Build with Maven
