+++ IF/src/main/java/com/github/stefvanschie/inventoryframework/util/version/VersionMatcher.java
@@ -299,8 +299,8 @@ public class VersionMatcher {
             com.github.stefvanschie.inventoryframework.nms.v1_21_4.AnvilInventoryImpl.class);
         ANVIL_INVENTORIES.put(Version.V1_21_5,
             com.github.stefvanschie.inventoryframework.nms.v1_21_5.AnvilInventoryImpl.class);
-        ANVIL_INVENTORIES.put(Version.V1_21_6_7,
-            com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.AnvilInventoryImpl.class);
+        ANVIL_INVENTORIES.put(Version.V1_21_6_8,
+            com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.AnvilInventoryImpl.class);
 
         BEACON_INVENTORIES = new EnumMap<>(Version.class);
         BEACON_INVENTORIES.put(Version.V1_16_1,
@@ -351,8 +351,8 @@ public class VersionMatcher {
             com.github.stefvanschie.inventoryframework.nms.v1_21_4.BeaconInventoryImpl.class);
         BEACON_INVENTORIES.put(Version.V1_21_5,
             com.github.stefvanschie.inventoryframework.nms.v1_21_5.BeaconInventoryImpl.class);
-        BEACON_INVENTORIES.put(Version.V1_21_6_7,
-            com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.BeaconInventoryImpl.class);
+        BEACON_INVENTORIES.put(Version.V1_21_6_8,
+            com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.BeaconInventoryImpl.class);
 
         CARTOGRAPHY_TABLE_INVENTORIES = new EnumMap<>(Version.class);
         CARTOGRAPHY_TABLE_INVENTORIES.put(Version.V1_16_1,
@@ -403,8 +403,8 @@ public class VersionMatcher {
             com.github.stefvanschie.inventoryframework.nms.v1_21_4.CartographyTableInventoryImpl.class);
         CARTOGRAPHY_TABLE_INVENTORIES.put(Version.V1_21_5,
             com.github.stefvanschie.inventoryframework.nms.v1_21_5.CartographyTableInventoryImpl.class);
-        CARTOGRAPHY_TABLE_INVENTORIES.put(Version.V1_21_6_7,
-            com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.CartographyTableInventoryImpl.class);
+        CARTOGRAPHY_TABLE_INVENTORIES.put(Version.V1_21_6_8,
+            com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.CartographyTableInventoryImpl.class);
 
         ENCHANTING_TABLE_INVENTORIES = new EnumMap<>(Version.class);
         ENCHANTING_TABLE_INVENTORIES.put(Version.V1_16_1,
@@ -455,8 +455,8 @@ public class VersionMatcher {
             com.github.stefvanschie.inventoryframework.nms.v1_21_4.EnchantingTableInventoryImpl.class);
         ENCHANTING_TABLE_INVENTORIES.put(Version.V1_21_5,
             com.github.stefvanschie.inventoryframework.nms.v1_21_5.EnchantingTableInventoryImpl.class);
-        ENCHANTING_TABLE_INVENTORIES.put(Version.V1_21_6_7,
-            com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.EnchantingTableInventoryImpl.class);
+        ENCHANTING_TABLE_INVENTORIES.put(Version.V1_21_6_8,
+            com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.EnchantingTableInventoryImpl.class);
 
         GRINDSTONE_INVENTORIES = new EnumMap<>(Version.class);
         GRINDSTONE_INVENTORIES.put(Version.V1_16_1,
@@ -507,8 +507,8 @@ public class VersionMatcher {
             com.github.stefvanschie.inventoryframework.nms.v1_21_4.GrindstoneInventoryImpl.class);
         GRINDSTONE_INVENTORIES.put(Version.V1_21_5,
             com.github.stefvanschie.inventoryframework.nms.v1_21_5.GrindstoneInventoryImpl.class);
-        GRINDSTONE_INVENTORIES.put(Version.V1_21_6_7,
-            com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.GrindstoneInventoryImpl.class);
+        GRINDSTONE_INVENTORIES.put(Version.V1_21_6_8,
+            com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.GrindstoneInventoryImpl.class);
 
         MERCHANT_INVENTORIES = new EnumMap<>(Version.class);
         MERCHANT_INVENTORIES.put(Version.V1_16_1,
@@ -559,8 +559,8 @@ public class VersionMatcher {
             com.github.stefvanschie.inventoryframework.nms.v1_21_4.MerchantInventoryImpl.class);
         MERCHANT_INVENTORIES.put(Version.V1_21_5,
             com.github.stefvanschie.inventoryframework.nms.v1_21_5.MerchantInventoryImpl.class);
-        MERCHANT_INVENTORIES.put(Version.V1_21_6_7,
-            com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.MerchantInventoryImpl.class);
+        MERCHANT_INVENTORIES.put(Version.V1_21_6_8,
+            com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.MerchantInventoryImpl.class);
 
         SMITHING_TABLE_INVENTORIES = new EnumMap<>(Version.class);
         SMITHING_TABLE_INVENTORIES.put(Version.V1_19_4,
@@ -587,8 +587,8 @@ public class VersionMatcher {
             com.github.stefvanschie.inventoryframework.nms.v1_21_4.SmithingTableInventoryImpl.class);
         SMITHING_TABLE_INVENTORIES.put(Version.V1_21_5,
             com.github.stefvanschie.inventoryframework.nms.v1_21_5.SmithingTableInventoryImpl.class);
-        SMITHING_TABLE_INVENTORIES.put(Version.V1_21_6_7,
-            com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.SmithingTableInventoryImpl.class);
+        SMITHING_TABLE_INVENTORIES.put(Version.V1_21_6_8,
+            com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.SmithingTableInventoryImpl.class);
 
         LEGACY_SMITHING_TABLE_INVENTORIES = new EnumMap<>(Version.class);
         LEGACY_SMITHING_TABLE_INVENTORIES.put(Version.V1_16_1,
@@ -667,7 +667,7 @@ public class VersionMatcher {
             com.github.stefvanschie.inventoryframework.nms.v1_21_4.StonecutterInventoryImpl.class);
         STONECUTTER_INVENTORIES.put(Version.V1_21_5,
             com.github.stefvanschie.inventoryframework.nms.v1_21_5.StonecutterInventoryImpl.class);
-        STONECUTTER_INVENTORIES.put(Version.V1_21_6_7,
-            com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.StonecutterInventoryImpl.class);
+        STONECUTTER_INVENTORIES.put(Version.V1_21_6_8,
+            com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.StonecutterInventoryImpl.class);
     }
 }
