+++ nms/1_21_6-8/src/main/java/com/github/stefvanschie/inventoryframework/nms/v1_21_6_8/SmithingTableInventoryImpl.java
@@ -1,8 +1,8 @@
-package com.github.stefvanschie.inventoryframework.nms.v1_21_6_7;
+package com.github.stefvanschie.inventoryframework.nms.v1_21_6_8;
 
-import com.github.stefvanschie.inventoryframework.abstraction.GrindstoneInventory;
+import com.github.stefvanschie.inventoryframework.abstraction.SmithingTableInventory;
 import com.github.stefvanschie.inventoryframework.adventuresupport.TextHolder;
-import com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.util.TextHolderUtil;
+import com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.util.TextHolderUtil;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.CompoundContainer;
@@ -10,11 +10,10 @@ import net.minecraft.world.Container;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.inventory.ContainerLevelAccess;
-import net.minecraft.world.inventory.GrindstoneMenu;
-import net.minecraft.world.inventory.Slot;
-import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftInventoryGrindstone;
+import net.minecraft.world.inventory.*;
+import net.minecraft.world.item.ItemStack;
+import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftInventory;
+import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftInventorySmithing;
 import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftInventoryView;
 import org.bukkit.entity.HumanEntity;
 import org.bukkit.event.inventory.InventoryType;
@@ -24,17 +23,17 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
- * Internal grindstone inventory for 1.21.6 and 1.21.7
+ * Internal smithing table inventory for 1.21.6 - 1.21.8. This is only available for Minecraft 1.20 and higher.
  *
  * @since 0.11.1
  */
-public class GrindstoneInventoryImpl extends GrindstoneInventory {
+public class SmithingTableInventoryImpl extends SmithingTableInventory {
 
     @NotNull
     @Contract(pure = true)
     @Override
     public Inventory createInventory(@NotNull TextHolder title) {
-        SimpleContainer resultSlot = new SimpleContainer(1);
+        ResultContainer resultSlot = new ResultContainer();
 
         Container container = new InventoryViewProvider() {
             @NotNull
@@ -45,7 +44,7 @@ public class GrindstoneInventoryImpl extends GrindstoneInventory {
                     @Nullable net.minecraft.world.entity.player.Inventory inventory,
                     @NotNull Player player
             ) {
-                return new ContainerGrindstoneImpl(containerId, player, this, resultSlot);
+                return new ContainerSmithingTableImpl(containerId, player, this, resultSlot);
             }
 
             @NotNull
@@ -56,12 +55,12 @@ public class GrindstoneInventoryImpl extends GrindstoneInventory {
             }
         };
 
-        return new CraftInventoryGrindstone(container, resultSlot) {
+        return new CraftInventorySmithing(null, container, resultSlot) {
             @NotNull
             @Contract(pure = true)
             @Override
             public InventoryType getType() {
-                return InventoryType.GRINDSTONE;
+                return InventoryType.SMITHING;
             }
 
             @Override
@@ -81,21 +80,21 @@ public class GrindstoneInventoryImpl extends GrindstoneInventory {
     private abstract static class InventoryViewProvider extends SimpleContainer implements MenuProvider {
 
         /**
-         * Creates a new inventory view provider with two slots.
+         * Creates a new inventory view provider with three slots.
          *
          * @since 0.11.1
          */
         public InventoryViewProvider() {
-            super(2);
+            super(3);
         }
     }
 
     /**
-     * A custom container grindstone.
+     * A custom container smithing table
      *
      * @since 0.11.1
      */
-    private static class ContainerGrindstoneImpl extends GrindstoneMenu {
+    private static class ContainerSmithingTableImpl extends SmithingMenu {
 
         /**
          * The human entity viewing this menu.
@@ -113,7 +112,7 @@ public class GrindstoneInventoryImpl extends GrindstoneInventory {
          * The container for the result slot.
          */
         @NotNull
-        private final SimpleContainer resultSlot;
+        private final ResultContainer resultSlot;
 
         /**
          * The corresponding Bukkit view. Will be not null after the first call to {@link #getBukkitView()} and null
@@ -123,7 +122,7 @@ public class GrindstoneInventoryImpl extends GrindstoneInventory {
         private CraftInventoryView<?, ?> bukkitEntity;
 
         /**
-         * Creates a new custom grindstone container for the specified player.
+         * Creates a new custom smithing table container for the specified player
          *
          * @param containerId the container id
          * @param player the player
@@ -131,11 +130,11 @@ public class GrindstoneInventoryImpl extends GrindstoneInventory {
          * @param resultSlot the result slot
          * @since 0.11.1
          */
-        public ContainerGrindstoneImpl(
+        public ContainerSmithingTableImpl(
                 int containerId,
                 @NotNull Player player,
                 @NotNull SimpleContainer itemsSlots,
-                @NotNull SimpleContainer resultSlot
+                @NotNull ResultContainer resultSlot
         ) {
             super(containerId, player.getInventory(), ContainerLevelAccess.create(player.level(), BlockPos.ZERO));
 
@@ -150,6 +149,7 @@ public class GrindstoneInventoryImpl extends GrindstoneInventory {
             updateSlot(0, container);
             updateSlot(1, container);
             updateSlot(2, container);
+            updateSlot(3, container);
         }
 
         @NotNull
@@ -159,21 +159,39 @@ public class GrindstoneInventoryImpl extends GrindstoneInventory {
                 return this.bukkitEntity;
             }
 
-            CraftInventoryGrindstone inventory = new CraftInventoryGrindstone(this.itemsSlots, this.resultSlot);
+            CraftInventory inventory = new CraftInventorySmithing(
+                    this.access.getLocation(),
+                    this.itemsSlots,
+                    this.resultSlot
+            );
 
             this.bukkitEntity = new CraftInventoryView<>(this.humanEntity, inventory, this);
 
             return this.bukkitEntity;
         }
 
+        @Contract(pure = true, value = "_ -> true")
         @Override
-        public void slotsChanged(@Nullable Container container) {}
+        public boolean stillValid(@Nullable net.minecraft.world.entity.player.Player nmsPlayer) {
+            return true;
+        }
 
         @Override
-        public void removed(@Nullable Player player) {}
+        public void slotsChanged(Container container) {}
 
         @Override
-        protected void clearContainer(@Nullable Player player, @Nullable Container container) {}
+        public void removed(net.minecraft.world.entity.player.Player nmsPlayer) {}
+
+        @Override
+        public void createResult() {}
+
+        @Override
+        protected void onTake(net.minecraft.world.entity.player.Player player, ItemStack stack) {}
+
+        @Override
+        protected boolean mayPickup(net.minecraft.world.entity.player.Player player, boolean present) {
+            return true;
+        }
 
         /**
          * Updates the current slot at the specified index to a new slot. The new slot will have the same slot, x, y,
