+++ nms/1_21_6-8/src/main/java/com/github/stefvanschie/inventoryframework/nms/v1_21_6_8/StonecutterInventoryImpl.java
@@ -1,20 +1,18 @@
-package com.github.stefvanschie.inventoryframework.nms.v1_21_6_7;
+package com.github.stefvanschie.inventoryframework.nms.v1_21_6_8;
 
-import com.github.stefvanschie.inventoryframework.abstraction.EnchantingTableInventory;
+import com.github.stefvanschie.inventoryframework.abstraction.StonecutterInventory;
 import com.github.stefvanschie.inventoryframework.adventuresupport.TextHolder;
-import com.github.stefvanschie.inventoryframework.nms.v1_21_6_7.util.TextHolderUtil;
+import com.github.stefvanschie.inventoryframework.nms.v1_21_6_8.util.TextHolderUtil;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
+import net.minecraft.world.CompoundContainer;
 import net.minecraft.world.Container;
 import net.minecraft.world.MenuProvider;
 import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.inventory.ContainerLevelAccess;
-import net.minecraft.world.inventory.EnchantmentMenu;
-import net.minecraft.world.inventory.Slot;
-import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftInventoryEnchanting;
-import org.bukkit.craftbukkit.v1_21_R5.inventory.view.CraftEnchantmentView;
+import net.minecraft.world.inventory.*;
+import org.bukkit.craftbukkit.v1_21_R5.inventory.CraftInventoryStonecutter;
+import org.bukkit.craftbukkit.v1_21_R5.inventory.view.CraftStonecutterView;
 import org.bukkit.entity.HumanEntity;
 import org.bukkit.event.inventory.InventoryType;
 import org.bukkit.inventory.Inventory;
@@ -22,12 +20,19 @@ import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class EnchantingTableInventoryImpl extends EnchantingTableInventory {
+/**
+ * Internal stonecutter inventory for 1.21.6 - 1.21.8
+ *
+ * @since 0.11.1
+ */
+public class StonecutterInventoryImpl extends StonecutterInventory {
 
     @NotNull
     @Contract(pure = true)
     @Override
     public Inventory createInventory(@NotNull TextHolder title) {
+        SimpleContainer resultSlot = new SimpleContainer(1);
+
         Container container = new InventoryViewProvider() {
             @NotNull
             @Contract(pure = true)
@@ -37,7 +42,7 @@ public class EnchantingTableInventoryImpl extends EnchantingTableInventory {
                     @Nullable net.minecraft.world.entity.player.Inventory inventory,
                     @NotNull Player player
             ) {
-                return new ContainerEnchantingTableImpl(containerId, player, this);
+                return new ContainerStonecutterImpl(containerId, player, this, resultSlot);
             }
 
             @NotNull
@@ -48,12 +53,12 @@ public class EnchantingTableInventoryImpl extends EnchantingTableInventory {
             }
         };
 
-        return new CraftInventoryEnchanting(container) {
+        return new CraftInventoryStonecutter(container, resultSlot) {
             @NotNull
             @Contract(pure = true)
             @Override
             public InventoryType getType() {
-                return InventoryType.ENCHANTING;
+                return InventoryType.STONECUTTER;
             }
 
             @Override
@@ -73,21 +78,21 @@ public class EnchantingTableInventoryImpl extends EnchantingTableInventory {
     private abstract static class InventoryViewProvider extends SimpleContainer implements MenuProvider {
 
         /**
-         * Creates a new inventory view provider with two slots.
+         * Creates a new inventory view provider with three slots.
          *
          * @since 0.11.1
          */
         public InventoryViewProvider() {
-            super(2);
+            super(1);
         }
     }
 
     /**
-     * A custom container enchanting table.
+     * A custom container stonecutter
      *
      * @since 0.11.1
      */
-    private static class ContainerEnchantingTableImpl extends EnchantmentMenu {
+    private static class ContainerStonecutterImpl extends StonecutterMenu {
 
         /**
          * The human entity viewing this menu.
@@ -96,64 +101,84 @@ public class EnchantingTableInventoryImpl extends EnchantingTableInventory {
         private final HumanEntity humanEntity;
 
         /**
-         * The container for the input slots.
+         * The container for the items slots.
          */
         @NotNull
-        private final SimpleContainer inputSlots;
+        private final SimpleContainer inputSlot;
+
+        /**
+         * The container for the result slot.
+         */
+        @NotNull
+        private final SimpleContainer resultSlot;
 
         /**
          * The corresponding Bukkit view. Will be not null after the first call to {@link #getBukkitView()} and null
          * prior.
          */
         @Nullable
-        private CraftEnchantmentView bukkitEntity;
+        private CraftStonecutterView bukkitEntity;
 
         /**
-         * Creates a new custom enchanting table container for the specified player.
+         * Creates a new custom stonecutter container for the specified player
          *
          * @param containerId the container id
          * @param player the player
-         * @param inputSlots the input slots
+         * @param inputSlot the input slot
+         * @param resultSlot the result slot
          * @since 0.11.1
          */
-        public ContainerEnchantingTableImpl(
+        public ContainerStonecutterImpl(
                 int containerId,
                 @NotNull Player player,
-                @NotNull SimpleContainer inputSlots
+                @NotNull SimpleContainer inputSlot,
+                @NotNull SimpleContainer resultSlot
         ) {
             super(containerId, player.getInventory(), ContainerLevelAccess.create(player.level(), BlockPos.ZERO));
 
             this.humanEntity = player.getBukkitEntity();
-            this.inputSlots = inputSlots;
+            this.inputSlot = inputSlot;
+            this.resultSlot = resultSlot;
 
             super.checkReachable = false;
 
-            updateSlot(0, inputSlots);
-            updateSlot(1, inputSlots);
+            CompoundContainer container = new CompoundContainer(inputSlot, resultSlot);
+
+            updateSlot(0, container);
+            updateSlot(1, container);
         }
 
         @NotNull
         @Override
-        public CraftEnchantmentView getBukkitView() {
+        public CraftStonecutterView getBukkitView() {
             if (this.bukkitEntity != null) {
                 return this.bukkitEntity;
             }
 
-            CraftInventoryEnchanting inventory = new CraftInventoryEnchanting(this.inputSlots);
+            CraftInventoryStonecutter inventory = new CraftInventoryStonecutter(this.inputSlot, this.resultSlot);
 
-            this.bukkitEntity = new CraftEnchantmentView(this.humanEntity, inventory, this);
+            this.bukkitEntity = new CraftStonecutterView(this.humanEntity, inventory, this);
 
             return this.bukkitEntity;
         }
 
+        @Contract(pure = true, value = "_ -> true")
         @Override
-        public void slotsChanged(@Nullable Container container) {}
+        public boolean stillValid(@Nullable net.minecraft.world.entity.player.Player nmsPlayer) {
+            return true;
+        }
 
         @Override
-        public void removed(@Nullable Player player) {}
+        public void slotsChanged(Container container) {}
 
         @Override
-        protected void clearContainer(@Nullable Player player, @Nullable Container container) {}
+        public void removed(net.minecraft.world.entity.player.Player nmsPlayer) {}
+
+        @Contract(value = "_, _ -> false", pure = true)
+        @Override
+        public boolean clickMenuButton(@Nullable Player player, int index) {
+            return false;
+        }
 
         /**
          * Updates the current slot at the specified index to a new slot. The new slot will have the same slot, x, y,
