+++ pom.xml
@@ -7,7 +7,7 @@
     <modules>
         <module>IF</module>
         <module>nms/abstraction</module>
-        <module>nms/1_21_6-7</module>
+        <module>nms/1_21_6-8</module>
         <module>nms/1_21_5</module>
         <module>nms/1_21_4</module>
         <module>nms/1_21_2-3</module>
